tests: move replica tests to nightly tests only - revert test changes for k3d / microk8s - add test_upload_replicas to nightly to test replicas after upload - add create replicas test to test_crawl_timeout - add delete replicas test to test_delete_crawls - add test_z_background_jobs to nightly - increase page limit, lower crawl stop timeout to avoid crawls finished too quickly! - add test-nightly-addons.yaml for configuring replicas
