+++ .github/workflows/k3d-nightly-ci.yaml
@@ -16,6 +16,7 @@ jobs:
           cluster-name: btrix-nightly
           args: >-
             -p "30870:30870@agent:0:direct"
+            -p "30090:30090@agent:0:direct"
             --agents 1
             --no-lb
             --k3s-arg "--disable=traefik,servicelb@server:*"
@@ -63,7 +64,7 @@ jobs:
 
       - name: Start Cluster with Helm
         run: |
-          helm upgrade --install -f ./chart/values.yaml -f ./chart/test/test.yaml --set invite_expire_seconds=10 --set max_pages_per_crawl=10 btrix ./chart/
+          helm upgrade --install -f ./chart/values.yaml -f ./chart/test/test.yaml -f ./chart/test/test-nightly-addons.yaml btrix ./chart/
 
       - name: Install Python
         uses: actions/setup-python@v3
