+++ .github/workflows/microk8s-ci.yaml
@@ -2,14 +2,14 @@ name: Cluster Run (MicroK8s)
 
 on:
   push:
-    #branches:
-    #  - main
-    #paths:
-    #  - 'backend/**'
-    #  - 'chart/**'
+    branches:
+      - main
+    paths:
+      - 'backend/**'
+      - 'chart/**'
 
 env:
-  HOST_URL_FROM_K8S: http://10.0.1.1
+  ECHO_SERVER_HOST_URL: http://10.0.1.1:18080
 
 jobs:
   btrix-microk8s-test:
@@ -62,9 +62,6 @@ jobs:
 
       - name: Wait for all pods to be ready
         run: sudo microk8s kubectl wait --for=condition=ready pod --all --timeout=240s
-
-      - name: Create Extra Test Buckets
-        run:  sudo microk8s kubectl exec -i deployment/local-minio -c minio mkdir /data/replica-0
 
       - name: Run Tests
         run: pytest -vv ./backend/test/test_*.py
