+++ backend/test/test_uploads.py
@@ -1,14 +1,9 @@
 import requests
 import os
-import time
 from urllib.parse import urljoin
 
 from .conftest import API_PREFIX
-from .utils import (
-    read_in_chunks,
-    verify_file_replicated,
-    verify_file_and_replica_deleted,
-)
+from .utils import read_in_chunks
 
 upload_id = None
 upload_id_2 = None
@@ -210,58 +205,6 @@ def test_get_upload_replay_json_admin(
     assert "files" not in data
 
 
-def test_upload_file_replicated(admin_auth_headers, default_org_id):
-    time.sleep(20)
-
-    # Verify replication job was successful
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/jobs?sortBy=started&sortDirection=-1&jobType=create-replica",
-        headers=admin_auth_headers,
-    )
-    assert r.status_code == 200
-    latest_job = r.json()["items"][0]
-    assert latest_job["type"] == "create-replica"
-    job_id = latest_job["id"]
-
-    attempts = 0
-    while attempts < 5:
-        r = requests.get(
-            f"{API_PREFIX}/orgs/{default_org_id}/jobs/{job_id}",
-            headers=admin_auth_headers,
-        )
-        assert r.status_code == 200
-        job = r.json()
-        finished = latest_job.get("finished")
-        if not finished:
-            attempts += 1
-            time.sleep(10)
-            continue
-
-        assert job["success"]
-        break
-
-    # Verify file updated
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/uploads/{upload_id}/replay.json",
-        headers=admin_auth_headers,
-    )
-    assert r.status_code == 200
-    data = r.json()
-    files = data.get("resources")
-    assert files
-    for file_ in files:
-        assert file_["numReplicas"] == 1
-
-    # Verify replica is stored
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/jobs/{job_id}", headers=admin_auth_headers
-    )
-    assert r.status_code == 200
-    job = r.json()
-    print(job["file_path"])
-    verify_file_replicated(job["file_path"])
-
-
 def test_replace_upload(admin_auth_headers, default_org_id, uploads_collection_id):
     actual_id = do_upload_replace(
         admin_auth_headers, default_org_id, upload_id, uploads_collection_id
@@ -390,45 +333,6 @@ def test_ensure_deleted(admin_auth_headers, default_org_id):
     for res in results["items"]:
         if res["id"] == upload_id:
             assert False
-
-    time.sleep(20)
-
-    # Verify delete replica job was successful
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/jobs?sortBy=started&sortDirection=-1&jobType=delete-replica",
-        headers=admin_auth_headers,
-    )
-    assert r.status_code == 200
-    latest_job = r.json()["items"][0]
-    assert latest_job["type"] == "delete-replica"
-    job_id = latest_job["id"]
-
-    attempts = 0
-    while attempts < 5:
-        r = requests.get(
-            f"{API_PREFIX}/orgs/{default_org_id}/jobs/{job_id}",
-            headers=admin_auth_headers,
-        )
-        assert r.status_code == 200
-        job = r.json()
-        finished = latest_job.get("finished")
-        if not finished:
-            attempts += 1
-            time.sleep(10)
-            continue
-
-        assert job["success"]
-        break
-
-    time.sleep(10)
-
-    # Verify replica is no longer stored
-    r = requests.get(
-        f"{API_PREFIX}/orgs/{default_org_id}/jobs/{job_id}", headers=admin_auth_headers
-    )
-    assert r.status_code == 200
-    job = r.json()
-    verify_file_and_replica_deleted(job["file_path"])
 
 
 def test_replace_upload_non_existent(
