+++ backend/test_nightly/test_delete_crawls.py
@@ -3,6 +3,7 @@ import requests
 import time
 
 from .conftest import API_PREFIX, HOST_PREFIX
+from .utils import verify_file_and_replica_deleted
 
 
 def test_delete_crawls(
@@ -86,3 +87,44 @@ def test_delete_crawls(
     # Verify that files are no longer retrievable from storage
     for url in crawl_resource_urls:
         assert not _file_is_retrievable(url)
+
+
+def test_delete_replica_job_run(admin_auth_headers, default_org_id):
+    time.sleep(20)
+
+    # Verify delete replica job was successful
+    r = requests.get(
+        f"{API_PREFIX}/orgs/{default_org_id}/jobs?sortBy=started&sortDirection=-1&jobType=delete-replica",
+        headers=admin_auth_headers,
+    )
+    assert r.status_code == 200
+    latest_job = r.json()["items"][0]
+    assert latest_job["type"] == "delete-replica"
+    job_id = latest_job["id"]
+
+    attempts = 0
+    while attempts < 5:
+        r = requests.get(
+            f"{API_PREFIX}/orgs/{default_org_id}/jobs/{job_id}",
+            headers=admin_auth_headers,
+        )
+        assert r.status_code == 200
+        job = r.json()
+        finished = latest_job.get("finished")
+        if not finished:
+            attempts += 1
+            time.sleep(10)
+            continue
+
+        assert job["success"]
+        break
+
+    time.sleep(10)
+
+    # Verify replica is no longer stored
+    r = requests.get(
+        f"{API_PREFIX}/orgs/{default_org_id}/jobs/{job_id}", headers=admin_auth_headers
+    )
+    assert r.status_code == 200
+    job = r.json()
+    verify_file_and_replica_deleted(job["file_path"])
